requirements_cuda.txt
# CUDA-optimized requirements for LightGlue scene matching
# Install with: uv pip install -r requirements_cuda.txt

# Core ML libraries
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Computer vision and image processing
opencv-python>=4.8.0
pillow>=10.0.0

# Scientific computing
numpy>=1.24.0
scikit-learn>=1.3.0

# Data handling
h5py>=3.9.0
pandas>=2.0.0

# Visualization
matplotlib>=3.7.0

# Progress bars and utilities
tqdm>=4.65.0

# LightGlue (install separately)
# git clone https://github.com/cvg/LightGlue.git && pip install -e ./LightGlue

# Optional: hloc for comparison
# git clone --recursive https://github.com/cvg/Hierarchical-Localization/ && pip install -e ./Hierarchical-Localization